+++ src/components/NavBar.jsx
@@ -23,7 +23,7 @@ const NavBar = () => {
   };
 
   const handleClickLogo = () => {
-    router.push("/App");
+    router.push("/");
   };
 
   const toggleDropdown = () => {
